+++ packages/viz/src/utils/data.js
@@ -384,6 +384,14 @@ class DataUtil {
   }
 
   getLatestPump = () => {
+    const pumpSettings = this.sort.byDate(this.filter.byType('pumpSettings').top(Infinity))[0]
+    if (pumpSettings?.payload?.device?.name && pumpSettings?.payload?.device?.manufacturer) {
+      return {
+        deviceModel: pumpSettings.payload.device.name,
+        manufacturer: pumpSettings.payload.device.manufacturer
+      }
+    }
+    /*If no pumpSettings is found, we can fall back to old upload object*/
     const uploadData = this.sort.byDate(this.filter.byType('upload').top(Infinity))
     const latestPumpUpload = getLatestPumpUpload(uploadData)
     const latestUploadSource = _.get(latestPumpUpload, 'source', '').toLowerCase()