+++ packages/yourloops/app/app.tsx
@@ -47,7 +47,7 @@ const Yourloops = (): JSX.Element => {
       clientId={appConfig.AUTH0_CLIENT_ID}
       redirectUri={window.location.origin}
       useRefreshTokens
-      audience="https://api-ext.your-loops.com"
+      audience={`https://${appConfig.AUTH0_DOMAIN}/api/v2/`}
     >
       <BrowserRouter>
         <MetricsLocationListener />