+++ packages/yourloops/test/unit/jest.config.js
@@ -6,6 +6,6 @@ module.exports = {
 
   // The glob patterns Jest uses to detect test files
   testMatch: [
-    '<rootDir>/**/*.test.tsx'
+    '<rootDir>/**/*.test.*'
   ]
 }