+++ CHANGELOG.md
@@ -2,16 +2,18 @@
 Blip is the web front end for YourLoops system.
 It is based on Tidepool Blip 1.27.
 
-## 3.0.2-beta23 - 2022-11-17
+## 3.0.2-beta26 - 2022-11-17
 ### Added
 - YLP-1679 Allow caregivers to remove patients from their direct share
 - YLP-1850 Patient search should comply to INS standard
 ### Engineering use
 - YLP-1624 Fix error messages when inviting a patient already in the team
 - YLP-1680 getPatientData Api method has a wrong parameter
+- YLP-1726 Rework signup with Auth0
 - YLP-1805 Push error to the backend
 - YLP-1873 Fix patient profile update not visible in dashboard
 - YLP-1874 Fix PDF report generation
+- YLP-1894 Retrieve pump settings in pump settings object if no upload object available
 - YLP-1892 Creation of an integration test for the team form
 - YLP-1901 CSV generation is not using the correct time frame
 - YLP-1903 Timezone detection should exclude some data types